+++ exasol_transformers_extension/udfs/models/prediction_tasks/prediction_task.py
@@ -37,7 +37,7 @@ class PredictionTask(Protocol):
         self, model_df: pd.DataFrame
     ) -> list[
         pd.DataFrame
-    ]:  # todo can probs use generators as output if also use them as input. better?
+    ]:
         """
         `extract_unique_param_based_dataframes` : Even if the data in a given
         dataframe all have the same model, there might be differences within the given
